requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
twilio==9.0.2
razorpay==1.4.1
google-generativeai==0.4.1
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
fpdf2==2.7.9
